+++ NEWS.md
@@ -1,3 +1,8 @@
+# documentNumbering 0.0.2
+
+* Added an alternative way to do things, {dn_counter}, an R6 object that you
+  call on itself for side effects.
+
 # documentNumbering 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.